Move PropertyEvaluationContext to a separate file
CURA-4186
